+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/SignalingChannel.kt
@@ -25,7 +25,7 @@ interface SignalingChannel {
         fun onConnect()
         fun onDisconnect()
         fun onInitialOffer(offerMessage: OfferMessage)
-        fun onSwitch(switchMessage: SwitchMessage)
+        fun onSwitched(switchedMessage: SwitchedMessage)
         fun onUpdatedOffer(sdp: String)
         fun onReOffer(sdp: String)
         fun onError(reason: SoraErrorReason)
@@ -189,7 +189,7 @@ class SignalingChannelImpl @JvmOverloads constructor(
         val switchMessage = MessageConverter.parseSwitchMessage(text)
         SoraLogger.d(TAG, "[signaling:$role] <- switch ${switchMessage}")
 
-        listener?.onSwitch(switchMessage)
+        listener?.onSwitched(switchMessage)
     }
 
     private fun onUpdateMessage(text: String) {
@@ -296,7 +296,7 @@ class SignalingChannelImpl @JvmOverloads constructor(
                     MessageConverter.parseType(json)?.let {
                         when (it) {
                             "offer"    -> onOfferMessage(json)
-                            "switch"   -> onSwitchMessage(json)
+                            "switched" -> onSwitchMessage(json)
                             "ping"     -> onPingMessage(json)
                             "update"   -> onUpdateMessage(json)
                             "re-offer" -> onReOfferMessage(json)