+++ CHANGES.md
@@ -15,7 +15,7 @@
   - data_channel_signlaing, ignore_disconnect_websocket パラメータ設定を追加する
   - onDataChannel コールバックを実装する
   - 各 label に対応するデータチャネル関係のコールバックを実装する
-  - WebSocket 側の `type:switch` 受信の処理を追加する
+  - WebSocket 側の `type:switched` 受信の処理を追加する
   - @shino
 - [FIX] 終了前にシグナリング Disconnect メッセージ送信を追加する
   - 状態により WebSocket, DataChannel どちらかで送信する